requirements.txt
psycopg2-binary
python-dotenv
rich
streamlit
pandas
plotly